# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.CreateRunRequest:
    properties:
      jobId:
        type: string
      params:
        additionalProperties: true
        type: object
    type: object
  api.CreateRunResponse:
    properties:
      id:
        type: string
    type: object
...
  api.Run:
    properties:
      completedAt:
        type: string
      createdAt:
        type: string
      error:
        type: string
      id:
        type: string
      jobId:
        type: string
      params:
        additionalProperties: true
        type: object
      startedAt:
        type: string
      status:
        $ref: '#/definitions/api.RunStatus'
      updatedAt:
        type: string
    type: object
  api.RunListResponse:
    properties:
      cursor:
        type: string
      hasMore:
        type: boolean
      runs:
        items:
          $ref: '#/definitions/api.Run'
        type: array
    type: object
  api.RunStatus:
    enum:
    - pending
    - running
    - succeeded
    - failed
    - cancelled
    type: string
    x-enum-varnames:
    - RunStatusPending
    - RunStatusRunning
    - RunStatusSucceeded
    - RunStatusFailed
    - RunStatusCancelled
...
  api.ValidateWorkflowRequest:
    properties:
      input:
        additionalProperties: true
        type: object
      workflow:
        type: string
    type: object
  api.ValidateWorkflowResponse:
    properties:
      errors:
        items:
          type: string
        type: array
      valid:
        type: boolean
    type: object
...
  api.Workflow:
    properties:
      description:
        type: string
      name:
        type: string
      schema:
        additionalProperties: true
        type: object
      version:
        type: string
    type: object
...
host: localhost:3333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for the Agent Project Manager daemon. Orchestrates agent-based
    LLM workflows.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Agent Project Manager API
  version: "1.0"
paths:
...
  /runs:
    get:
      consumes:
      - application/json
      description: Get a paginated list of runs with optional filtering
      parameters:
      - description: Maximum number of runs to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Filter by job ID
        in: query
        name: jobId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.RunListResponse'
      summary: List runs
      tags:
      - runs
    post:
      consumes:
      - application/json
      description: Start a new run for a job
      parameters:
      - description: Run creation request
        in: body
        name: run
        required: true
        schema:
          $ref: '#/definitions/api.CreateRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.CreateRunResponse'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Create a new run
      tags:
      - runs
  /runs/{runId}:
    delete:
      consumes:
      - application/json
      description: Cancel a running or pending run
      parameters:
      - description: Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "404":
          description: Run not found
          schema:
            type: string
      summary: Cancel a run
      tags:
      - runs
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific run
      parameters:
      - description: Run ID
        in: path
        name: runId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Run'
        "404":
          description: Run not found
          schema:
            type: string
      summary: Get run details
      tags:
      - runs
...
  /workflows/{name}:
    get:
      consumes:
      - application/json
      description: Get schema and metadata for a specific workflow
      parameters:
      - description: Workflow name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Workflow'
        "404":
          description: Workflow not found
          schema:
            type: string
      summary: Get workflow details
      tags:
      - workflows
  /workflows/validate:
    post:
      consumes:
      - application/json
      description: Validate input parameters for a workflow
      parameters:
      - description: Workflow validation request
        in: body
        name: workflow
        required: true
        schema:
          $ref: '#/definitions/api.ValidateWorkflowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ValidateWorkflowResponse'
        "400":
          description: Invalid request
          schema:
            type: string
      summary: Validate workflow input
      tags:
      - workflows
schemes:
- http
- https
swagger: "2.0"