# internal/api/handlers_runs.go
package api

import (
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// handleCreateRun handles POST /runs
// @Summary      Create a new run
// @Description  Start a new run for a job
// @Tags         runs
// @Accept       json
// @Produce      json
// @Param        run   body      CreateRunRequest  true  "Run creation request"
// @Success      201   {object}  CreateRunResponse
// @Failure      400   {string}  string  "Invalid request"
// @Router       /runs [post]
func handleCreateRun(w http.ResponseWriter, r *http.Request) {
	var req CreateRunRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement run creation logic
	response := CreateRunResponse{
		ID: generateID(),
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(response)
}

// handleListRuns handles GET /runs
// @Summary      List runs
// @Description  Get a paginated list of runs with optional filtering
// @Tags         runs
// @Accept       json
// @Produce      json
// @Param        limit   query     int     false  "Maximum number of runs to return"
// @Param        cursor  query     string  false  "Cursor for pagination"
// @Param        status  query     string  false  "Filter by status"
// @Param        jobId   query     string  false  "Filter by job ID"
// @Success      200     {object}  RunListResponse
// @Router       /runs [get]
func handleListRuns(w http.ResponseWriter, r *http.Request) {
	// TODO: Parse query parameters: limit, cursor, status, jobId
	// limit := r.URL.Query().Get("limit")
	// cursor := r.URL.Query().Get("cursor")
	// status := r.URL.Query().Get("status")
	// jobId := r.URL.Query().Get("jobId")

	// TODO: Implement run listing logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode([]interface{}{})
}

// handleGetRun handles GET /runs/{runId}
// @Summary      Get run details
// @Description  Get detailed information about a specific run
// @Tags         runs
// @Accept       json
// @Produce      json
// @Param        runId   path      string  true  "Run ID"
// @Success      200     {object}  Run
// @Failure      404     {string}  string  "Run not found"
// @Router       /runs/{runId} [get]
func handleGetRun(w http.ResponseWriter, r *http.Request) {
	runID := chi.URLParam(r, "runId")

	// TODO: Implement run retrieval logic
	_ = runID

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{})
}

// handleDeleteRun handles DELETE /runs/{runId}
// @Summary      Cancel a run
// @Description  Cancel a running or pending run
// @Tags         runs
// @Accept       json
// @Produce      json
// @Param        runId   path      string  true  "Run ID"
// @Success      202     {string}  string  "Accepted"
// @Failure      404     {string}  string  "Run not found"
// @Router       /runs/{runId} [delete]
func handleDeleteRun(w http.ResponseWriter, r *http.Request) {
	runID := chi.URLParam(r, "runId")

	// TODO: Implement run cancellation logic
	_ = runID

	w.WriteHeader(http.StatusAccepted)
}