# internal/api/handlers_workflows.go
package api

import (
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// handleListWorkflows handles GET /workflows
// @Summary      List workflows
// @Description  Get a list of all available workflows
// @Tags         workflows
// @Accept       json
// @Produce      json
// @Success      200  {object}  WorkflowListResponse
// @Router       /workflows [get]
func handleListWorkflows(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement workflow listing logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode([]interface{}{})
}

// handleGetWorkflow handles GET /workflows/{name}
// @Summary      Get workflow details
// @Description  Get schema and metadata for a specific workflow
// @Tags         workflows
// @Accept       json
// @Produce      json
// @Param        name   path      string  true  "Workflow name"
// @Success      200    {object}  Workflow
// @Failure      404    {string}  string  "Workflow not found"
// @Router       /workflows/{name} [get]
func handleGetWorkflow(w http.ResponseWriter, r *http.Request) {
	name := chi.URLParam(r, "name")

	// TODO: Implement workflow schema/metadata retrieval
	_ = name

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{})
}

// handleValidateWorkflow handles POST /workflows/validate
// @Summary      Validate workflow input
// @Description  Validate input parameters for a workflow
// @Tags         workflows
// @Accept       json
// @Produce      json
// @Param        workflow  body      ValidateWorkflowRequest  true  "Workflow validation request"
// @Success      200       {object}  ValidateWorkflowResponse
// @Failure      400       {string}  string  "Invalid request"
// @Router       /workflows/validate [post]
func handleValidateWorkflow(w http.ResponseWriter, r *http.Request) {
	var req ValidateWorkflowRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement workflow validation logic
	response := ValidateWorkflowResponse{
		Valid:  true,
		Errors: []string{},
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}